LPS-179751 Extract ScreenNavigationEntry methods from CategoryDetailsScreenNavigationCategory to CategoryDetailsScreenNavigationEntry and return getCategoryKey in getEntryKey method
